Clustering/requirements.txt
# Core
pandas>=2.1.0
numpy>=1.26.0
lxml>=5.1.0
pyyaml>=6.0.0

# ML & Feature Engineering
scikit-learn>=1.4.0
scipy>=1.12.0

# Visualization
plotly>=5.18.0
kaleido>=0.2.1

# Dashboard
streamlit>=1.31.0

# Notebook
nbformat>=4.2.0

# Testing
pytest>=8.0.0
pytest-cov>=4.1.0

# Data I/O
pyarrow>=15.0.0
